services/agent/agent/requirements.txt
authlib>=1.3.0
cachetools>=5.3.0
fastapi>=0.111.0
uvicorn>=0.29.0
hvac>=2.1.0
psycopg2-binary>=2.9.9
structlog>=24.1.0
httpx>=0.27.0
